requirements.txt
click==8.1.3
colorama==0.4.6
flask-mysql-connector==1.1.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysql-connector==2.2.9
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
pip==22.3.1
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
setuptools==65.5.1
six==1.16.0
Werkzeug==2.2.3
wheel==0.38.4
